#0: Refactore flash decode kernel for future scaling
